Add files via upload
In this, we learn how to use the Scikit-Learn library to implement Simple linear regression. We downloaded a dataset that is related to fuel consumption and Carbon dioxide emission of cars. Then, we split our data into training and test sets, create a model using a training set, evaluate your model using a test set, and finally use the model to predict unknown values.